change String::print()

+++ OOPython/OOPython/Block.cpp
@@ -185,7 +185,7 @@ int pyBreakBlock::work(int workStatus, Varmap& varmap) {
 }
 
 int pyAssignBlock::work(int workStatus, Varmap& varmap) {
-	int l = beAssigned.size();
+	int l = (int)beAssigned.size();
 	for (int i = 0; i < l; i++) {
 		const string& s = (dynamic_cast<pyVariable*> (beAssigned[i]))->getName();
 		pyObjectPtr back = assigner[i]->work(varmap);
